update : removed taglinecard and renamed files to resumecard
